identity_mirror_agent/backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.3
sqlalchemy==2.0.32
python-dotenv==1.0.1
pydantic==2.8.2
email-validator==2.2.0
anthropic==0.34.2
textblob==0.18.0.post0
httpx==0.27.0
pytest==8.3.2
pytest-asyncio==0.23.8
psycopg